🔐 Verify ownership (#105)
