--- app/repo/page.tsx
@@ -1,14 +1,0 @@
-import BodyContent from "@/components/BodyContent";
-import ClientOnly from "@/components/ClientOnly";
-
-type Props = {};
-
-function RepositoryPage({}: Props) {
-  return (
-    <ClientOnly>
-      <BodyContent />
-    </ClientOnly>
-  );
-}
-
-export default RepositoryPage;
